+++ test/lib/next-test-utils.ts
@@ -26,6 +26,10 @@ import type { RequestInit, Response } from 'node-fetch'
 import type { NextServer } from 'next/dist/server/next'
 import type { BrowserInterface } from './browsers/base'
 
+import { shouldRunTurboDevTest } from './turbo'
+
+export { shouldRunTurboDevTest }
+
 export const nextServer = server
 export const pkg = _pkg
 
@@ -147,15 +151,6 @@ export function withQuery(
   return `${pathname}?${querystring}`
 }
 
-export function renderViaHTTP(
-  appPort: string | number,
-  pathname: string,
-  query?: Record<string, any> | string | undefined,
-  opts?: RequestInit
-) {
-  return fetchViaHTTP(appPort, pathname, query, opts).then((res) => res.text())
-}
-
 export function fetchViaHTTP(
   appPort: string | number,
   pathname: string,
@@ -176,6 +171,15 @@ export function fetchViaHTTP(
     },
     ...opts,
   })
+}
+
+export function renderViaHTTP(
+  appPort: string | number,
+  pathname: string,
+  query?: Record<string, any> | string | undefined,
+  opts?: RequestInit
+) {
+  return fetchViaHTTP(appPort, pathname, query, opts).then((res) => res.text())
 }
 
 export function findPort() {
@@ -998,32 +1002,6 @@ export function findAllTelemetryEvents(output: string, eventName: string) {
     JSON.parse(entry.pop())
   )
   return events.filter((e) => e.eventName === eventName).map((e) => e.payload)
-}
-
-/**
- * Utility function to determine if a given test case needs to run with --turbo.
- *
- * This is primarily for the gradual test enablement with latest turbopack upstream changes.
- *
- * Note: it could be possible to dynamically create test cases itself (createDevTest(): it.each([...])), but
- * it makes hard to conform with existing lint rules. Instead, starting off from manual fixture setup and
- * update test cases accordingly as turbopack changes enable more test cases.
- */
-export function shouldRunTurboDevTest(): boolean {
-  if (!!process.env.TEST_WASM) {
-    return false
-  }
-
-  const shouldRunTurboDev = !!process.env.TURBOPACK
-  // short-circuit to run all the test with --turbo enabled skips glob matching costs
-  if (shouldRunTurboDev) {
-    console.log(
-      `Running tests with turbopack via custom environment variable TURBOPACK`
-    )
-    return true
-  }
-
-  return false
 }
 
 type TestVariants = 'default' | 'turbo'
